Split the code into platform specfic modules
